Convert Clusters table actions to three-dot menu and add Permissions editing
